feat: automated tox testing & typed default params
