+++ tox.ini
@@ -1,7 +1,12 @@
 [tox]
-minversion = 3.8.0
-envlist = py38, py39, py310, flake8, mypy
+minversion = 3.9.0
+envlist = py39, py310, flake8, mypy
 isolated_build = true
+
+[gh-actions]
+python = 
+    3.9: py39
+    3.10: py310, flake8, mypy
 
 [testenv]
 setenv = 
@@ -23,6 +28,6 @@ commands =
 basepython = python3.10
 deps =
     -r{toxinidir}/requirements.txt
-    -r{toxinidir}/requirements_dev.txt
+    -r{toxinidir}/requirements-dev.txt
 commands =
-    mypy . --exclude=tests
+    mypy . --exclude=build
